+++ lib/em/gserver/base.rb
@@ -15,11 +15,12 @@ module EventMachine
         class Base
             include EventMachine::GServer::Utils
             include EventMachine::GServer::Listeners::Manager
+            include EasyDaemons::Helper
             
             DEFAULT_STOP_TIMEOUT               = 5.0 # seconds
             DEFAULT_HEARTBEAT_TIMEOUT          = 2.0 # seconds
             
-            attr_reader :logger, :listeners, :stopping, :stop_timeout,
+            attr_reader :name, :pid_file_path, :pid, :logger, :listeners, :stopping, :stop_timeout,
                 :heartbeat_interval, :status
             
             #
@@ -29,7 +30,8 @@ module EventMachine
             # @option opts [Float] :heartbeat_timeout (2) How often to check for dead connections.
             # @option opts [Array] :listeners An array of instances of subclasses of {EventMachine::GServer::Listeners::Base}. 
             #
-            def initialize(opts={})
+            def initialize(name, opts={})
+                @name = name
                 set_opts(opts)
                 set_listeners(opts)
                 @status = STOPPED_SYM
@@ -48,10 +50,10 @@ module EventMachine
             # Starts the server.
             #
             # Note that this is an idempotent operation. That is, you can call
-            # it several times. If the server is not running, it will start it.
+            # it several times. If the server is not running, it will run it.
             # Otherwise it will just return the current (:running) status.
             #
-            def start
+            def run
                 return @status unless @status == STOPPED_SYM
                 @status = STARTING_SYM
                 register_error_handler
@@ -64,6 +66,24 @@ module EventMachine
                     log(:info, "EventMachine::GServer up and running with pid=#{Process.pid}.")
                     @status = RUNNING_SYM
                 end
+            end
+            
+            #
+            # This method simply runs the run method as a daemon background process.
+            #
+            def start
+                if is_already_running?(@name)
+                    log(:info, "Process #{@name} is already running.")
+                    exit(1)
+                end
+                log(:info, "Starting #{@name} worker in the background ...")
+                daemonize(@name)
+                @pid = curr_pid
+                log(:info, "#{@name} worker successfully started in the background (pid=#{@pid}).")
+                log(:info, "Refreshing #{@pid_file_path} file with #{@pid}.")
+                refresh_pid_file(@pid_file_path, @pid)
+                log(:info, "#{@pid_file_path} successfully created with #{@pid}.")
+                run
             end
 
             #
@@ -95,6 +115,8 @@ module EventMachine
                 
                 @heartbeat_timeout = opts[HEARTBEAT_TIMEOUT_SYM].to_f
                 @heartbeat_timeout = DEFAULT_HEARTBEAT_TIMEOUT if @heartbeat_timeout <= 0.0
+                
+                @pid_file_path = opts[:pid_file_path] || "pids/#{@name}.pid"
             end
             
             #
@@ -115,6 +137,7 @@ module EventMachine
                     Signal.trap(signal) do
                         t = Thread.new do
                             stop
+                            remove_pid_file(@pid_file_path)
                         end
                         t.join
                     end